Update to datePicker
There was a use case where the date could be cleared by the user and then it was saving as a null value, which turns into a default 1969 date. There was also another use case where letters and text could be entered and saved. This change now validates all use cases.